The Character Selection Update [BIG]
For some reason, I didn't choose to commit at all during this phase as I was willing to throw all my changes out the window as I wasn't sure I would be able to put my idea into the project of how to make this work. But I've managed to do it, and this is the commit of all the changes made. For some reason, last time I tried to revert a commit, things just went wrong for some reason so I didn't want to commit just for this Character Selection implementation.

- Added Other Prefabs to SpawnablePrefabs i.e. EllairaPrefab, MatthewPrefab, BlakePrefab

- Added necessary components to those prefabs i.e. NetworkIdentity, Network Animator etc so they are the same as my test prefab.

- Copied the components of the Character Selection scene, and added it to the main menu scene, where I made it a prefab

- Add button in lobby to choose to select character

- Add actions to Character Select buttons (player icons) to set their player to the wanted icon and added script to Character Select prefab that controls this

- Change the SpawnSystem prefabs to include all the character prefabs (now is an array of prefabs)

- Added changes to the NetworkManagerLobby, NetworkRoomPlayerLobby, NetworkGamePlayerLobby to allow the selection of character by the player